created a HighFrequencyMotionManager
