+++ tricorder Watch App/Manager/CoreMotionManager.swift
@@ -12,18 +12,24 @@ import OSLog
 actor CoreMotionManager {
     let eventManager = EventManager.shared
 
-    let motionManager = CMBatchedSensorManager()
+    var motionManager: MotionManager? = nil
 }
 
 extension CoreMotionManager {
-    func stopUpdates() {
+    func stopUpdates() async {
         Logger.shared.debug("MotinManager: stopUpdates called on Thread \(Thread.current)")
 
-        motionManager.stopAccelerometerUpdates()
-        motionManager.stopDeviceMotionUpdates()
+        guard let motionManager else {
+            Logger.shared.error("Tried to stop updates but motionManager is nil")
+            return
+        }
+
+        await motionManager.stopUpdates()
+
+        self.motionManager = nil
     }
 
-    func startUpdates(recordingStart: Date) throws {
+    func startUpdates(recordingStart: Date, settings: Settings?) async throws {
         Logger.shared.debug("MotinManager: startUpdates called on Thread \(Thread.current)")
 
         guard
@@ -33,53 +39,17 @@ extension CoreMotionManager {
             throw CoreMotionManagerError.notSupported
         }
 
-        motionManager.startAccelerometerUpdates(handler: {
-            @Sendable (valuesedData, error) in
+        let motionManager = HighFrequencyMotionManager(handleUpdate: handleUpdate)
 
-            Logger.shared.debug("called on Thread \(Thread.current)")
+        //
+        //        if let settings {
+        //            motionManager.accelerometerDataFrequency = settings.motionSensorRecodingRates[.acceleration] ?? .max
+        //            motionManager.deviceMotionDataFrequency = settings.motionSensorRecodingRates[.deviceMotion] ?? .max
+        //        }
 
-            if let error = error {
-                Logger.shared.error(
-                    "Error starting AccelerometerUpdates: \(error.localizedDescription)"
-                )
-                return
-            }
+        try await motionManager.startUpdates(recordingStart: recordingStart)
 
-            guard let valuesedData = valuesedData else {
-                Logger.shared.error(
-                    "Error starting AccelerometerUpdates: did not recive any data"
-                )
-                return
-            }
-            self.consumeAccelerometerUpdates(
-                valuesedData: valuesedData,
-                recordingStart: recordingStart
-            )
-        })
-
-        motionManager.startDeviceMotionUpdates(handler: {
-            @Sendable (valuesedData, error) in
-
-            Logger.shared.debug("called on Thread \(Thread.current)")
-
-            if let error = error {
-                Logger.shared.error(
-                    "Error starting DeviceMotionUpdate: \(error.localizedDescription)"
-                )
-                return
-            }
-
-            guard let valuesedData = valuesedData else {
-                Logger.shared.error(
-                    "Error starting DeviceMotionUpdate: did not recive any data"
-                )
-                return
-            }
-            self.consumeDeviceMotionUpdates(
-                valuesedData: valuesedData,
-                recordingStart: recordingStart
-            )
-        })
+        self.motionManager = motionManager
     }
 }
 
@@ -97,120 +67,6 @@ extension CoreMotionManager {
                 data: sensor
             ) as Void
         }
-    }
-
-    nonisolated func consumeDeviceMotionUpdates(
-        valuesedData: [CMDeviceMotion],
-        recordingStart: Date
-    ) {
-        Logger.shared.debug("called on Thread \(Thread.current)")
-
-        // todo make this more reusable
-        // todo do all of this in a different thread
-        var rotationRateValues: [MotionValue] = []
-        var userAccelerationValues: [MotionValue] = []
-        var gravityValues: [MotionValue] = []
-        var quaternionValues: [MotionValue] = []
-
-        valuesedData.forEach { data in
-            let dataDate = Date(
-                timeIntervalSince1970: data.timestamp.timeIntervalSince1970
-            )
-            rotationRateValues.append(
-                MotionValue(
-                    x: data.rotationRate.x,
-                    y: data.rotationRate.y,
-                    z: data.rotationRate.z,
-                    timestamp: dataDate
-                )
-            )
-            userAccelerationValues.append(
-                MotionValue(
-                    x: data.userAcceleration.x,
-                    y: data.userAcceleration.y,
-                    z: data.userAcceleration.z,
-                    timestamp: dataDate
-                )
-            )
-            gravityValues.append(
-                MotionValue(
-                    x: data.gravity.x,
-                    y: data.gravity.y,
-                    z: data.gravity.z,
-                    timestamp: dataDate
-                )
-            )
-            quaternionValues.append(
-                MotionValue(
-                    x: data.attitude.quaternion.x,
-                    y: data.attitude.quaternion.y,
-                    z: data.attitude.quaternion.z,
-                    w: data.attitude.quaternion.w,
-                    timestamp: dataDate
-                )
-            )
-        }
-
-        handleUpdate(
-            Sensor.motion(
-                .rotationRate,
-                recordingStartDate: recordingStart,
-                values: rotationRateValues
-            )
-        )
-        handleUpdate(
-            Sensor.motion(
-                .userAcceleration,
-                recordingStartDate: recordingStart,
-                values: userAccelerationValues
-            )
-        )
-        handleUpdate(
-            Sensor.motion(
-                .gravity,
-                recordingStartDate: recordingStart,
-                values: gravityValues
-            )
-        )
-        handleUpdate(
-            Sensor.motion(
-                .quaternion,
-                recordingStartDate: recordingStart,
-                values: quaternionValues
-            )
-        )
-    }
-
-    nonisolated func consumeAccelerometerUpdates(
-        valuesedData: [CMAccelerometerData],
-        recordingStart: Date
-    ) {
-        Logger.shared.debug("called on Thread \(Thread.current)")
-
-        var values: [MotionValue] = []
-
-        valuesedData.forEach { data in
-            values.append(
-                MotionValue(
-                    x: data.acceleration.x,
-                    y: data.acceleration.y,
-                    z: data.acceleration.z,
-
-                    // The timestamp is the amount of time in seconds since the device booted.
-                    timestamp: Date(
-                        timeIntervalSince1970: data.timestamp
-                            .timeIntervalSince1970
-                    )
-                )
-            )
-        }
-        handleUpdate(
-            Sensor.motion(
-                .acceleration,
-                recordingStartDate: recordingStart,
-                values: values
-            )
-        )
     }
 }
